+++ gdpr/src/main/java/com/kairos/response/dto/common/StorageFormatResponseDTO.java
@@ -9,7 +9,7 @@ import java.time.LocalDate;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class StorageFormatResponseDTO {
 
-    private Integer id;
+    private Long id;
 
     private String name;
 
@@ -19,7 +19,7 @@ public class StorageFormatResponseDTO {
 
     private LocalDate suggestedDate;
 
-    public StorageFormatResponseDTO(Integer id, String name, Long organizationId, SuggestedDataStatus suggestedDataStatus, LocalDate suggestedDate) {
+    public StorageFormatResponseDTO(Long id, String name, Long organizationId, SuggestedDataStatus suggestedDataStatus, LocalDate suggestedDate) {
         this.id = id;
         this.name = name;
         this.organizationId = organizationId;
@@ -39,9 +39,9 @@ public class StorageFormatResponseDTO {
 
     public void setOrganizationId(Long organizationId) { this.organizationId = organizationId; }
 
-    public Integer getId() { return id; }
+    public Long getId() { return id; }
 
-    public void setId(Integer id) { this.id = id; }
+    public void setId(Long id) { this.id = id; }
 
     public String getName() { return name; }
