+++ gdpr/src/main/java/com/kairos/response/dto/common/TechnicalSecurityMeasureResponseDTO.java
@@ -11,7 +11,7 @@ import java.time.LocalDate;
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
 public class TechnicalSecurityMeasureResponseDTO {
 
-    private Integer id;
+    private Long id;
 
     private String name;
 
@@ -21,7 +21,7 @@ public class TechnicalSecurityMeasureResponseDTO {
 
     private LocalDate suggestedDate;
 
-    public TechnicalSecurityMeasureResponseDTO(Integer id, String name, Long organizationId, SuggestedDataStatus suggestedDataStatus, LocalDate suggestedDate) {
+    public TechnicalSecurityMeasureResponseDTO(Long id, String name, Long organizationId, SuggestedDataStatus suggestedDataStatus, LocalDate suggestedDate) {
         this.id = id;
         this.name = name;
         this.organizationId = organizationId;
@@ -41,9 +41,9 @@ public class TechnicalSecurityMeasureResponseDTO {
 
     public void setOrganizationId(Long organizationId) { this.organizationId = organizationId; }
 
-    public Integer getId() { return id; }
+    public Long getId() { return id; }
 
-    public void setId(Integer id) { this.id = id; }
+    public void setId(Long id) { this.id = id; }
 
     public String getName() { return name; }
